feat: se modificaron las clases de los controladores sacados desde el archivo MercadoLibreController.php que después fue eliminado y además se cambiaron las rutas de la api.php
