+++ multistock-backend/app/Http/Controllers/MercadoLibre/Reportes/getSalesByMonthController.php
@@ -6,7 +6,7 @@ use App\Models\MercadoLibreCredential;
 use Illuminate\Support\Facades\Http;
 use Illuminate\Http\Request;
 
-class MercadoLibreDocumentsController extends Controller
+class getSalesByMonthController
 {
 
     public function getSalesByMonth($clientId)
