+++ multistock-backend/app/Http/Controllers/MercadoLibre/Reportes/summaryController.php
@@ -6,7 +6,7 @@ use App\Models\MercadoLibreCredential;
 use Illuminate\Support\Facades\Http;
 use Illuminate\Http\Request;
 
-class MercadoLibreDocumentsController extends Controller
+class summaryController
 {
 
 /**
